fix: smooth FAB transition by replacing AnimatedSwitcher with independent AnimatedOpacity per widget
Remove SmartFABWidget and SmartFABController in favour of two independently
positioned widgets (camera FAB right, selection pill left) each fading with
AnimatedOpacity. The previous AnimatedSwitcher approach caused a
position-jump and centre-snap artefact when deselecting photos.
